save bit ideas and cleaned Levi so now we have a fast but still somehat accyrate levi. tried to do something similar with Buro but proving to be difficult. Will look into making Buro better too but wont be too bummed if not

+++ Assets/Scripts/Utils/Objects.cs
@@ -49,7 +49,8 @@ public static class ZobristHashing
         {
             for (int j = 0; j < 12; j++)
             {
-                zobristTable[i, j] = (ulong)random.Next() << 32 | (ulong)random.Next();
+                // Cast random.Next() to uint before shifting
+                zobristTable[i, j] = ((ulong)(uint)random.Next() << 32) | (ulong)(uint)random.Next();
             }
         }
     }
@@ -108,4 +109,5 @@ public static class ZobristHashing
         return currentHash;
     }
 }
+
